backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv